// sdks/typescript/lilypad/generated/api/types/AnnotationCreate.ts
/**
 * This file was auto-generated by Fern from our API Definition.
 */

import * as Lilypad from "../index.js";

/**
 * Annotation create model.
 */
export interface AnnotationCreate {
    span_uuid?: string;
    project_uuid?: string;
    function_uuid?: string;
    label?: Lilypad.Label;
    reasoning?: string;
    type?: Lilypad.EvaluationType;
    data?: Record<string, unknown>;
    assigned_to?: string[];
    assignee_email?: string[];
}